requirements.txt
fastapi
uvicorn[standard]
strands-agents[ollama]
strands-agents-tools
python-dotenv
